Temporarily disable clippy unused_unit test
wasm-bindgen 0.2.79 introduced a change in code generation that makes clippy complain
A fix has already been merged to main (https://github.com/rustwasm/wasm-bindgen/pull/2778)
but we have to wait for the next official release to make use of it.
Once we update to the new wasm-bindgen with the fix we should remove
this workaround!